requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-SocketIO==5.3.6
bleach==6.0.0
email-validator==2.1.0
python-json-logger==2.0.7
validators==0.22.0
pydantic==2.5.0
pydantic-settings==2.1.0
virustotal-python==1.0.1
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.3.0
redis==5.0.0
gunicorn==21.2.0
gevent==23.9.1
